Remove opencl_device_selection.h
Always use OpenCL device selection if OpenCL is enabled.

This fixes a regression which was introduced by commit
5c6a57b727d7fd3b2e92a4a4123dd2e41907b6bd which removed
the definition for USE_DEVICE_SELECTION.
